+++ src/containers/TransactionPage/Offer/Offer.js
@@ -23,6 +23,7 @@ const Offer = props => {
     transactionRole,
     isNegotiationProcess,
     isRegularNegotiation,
+    isProviderBanned,
     intl,
   } = props;
 
@@ -33,6 +34,9 @@ const Offer = props => {
   const isProvider = transactionRole === 'provider';
   const protectedData = transaction?.attributes?.protectedData;
   const offers = transaction?.attributes?.metadata?.offers;
+  const providerDefaultMessage = !isProviderBanned
+    ? protectedData?.providerDefaultMessage
+    : intl.formatMessage({ id: 'TransactionPage.messageSenderBanned' });
 
   const processName = resolveLatestProcessName(transaction?.attributes?.processName);
   let process = null;
@@ -113,7 +117,7 @@ const Offer = props => {
       <TextMaybe
         heading={intl.formatMessage({ id: 'TransactionPage.Offer.providerDefaultMessageLabel' })}
         headingClassName={css.defaultMessageLabel}
-        text={protectedData?.providerDefaultMessage}
+        text={providerDefaultMessage}
         isOwn={isNegotiationProcess && isProvider}
         showText={isNegotiationProcess}
       />